+++ date-of-love - (intellij)/src/main/java/br/com/dateoflove/servlet/CancelarOrcamentoServlet.java
@@ -33,7 +33,7 @@ public class CancelarOrcamentoServlet extends HttpServlet {
             UsuarioDao usuarioDao = new UsuarioDao();
 
             Usuario usuario =  usuarioDao.buscarUsuarioPorId(idUsuario);
-            Orcamentos orcamentos = orcamentosDao.buscarOrcamentoPorId((long) idOrcamento);
+            Orcamentos orcamentos = orcamentosDao.buscarOrcamentoPorId(idOrcamento);
 
             Email email = new Email("","",usuario.getEmail(),"","","");
